+++ Ch08/pom.xml
@@ -80,6 +80,17 @@
     </pluginManagement>
   </build>
 
+  <dependencyManagement>
+    <dependencies>
+      <dependency>
+        <groupId>com.google.cloud</groupId>
+        <artifactId>libraries-bom</artifactId>
+        <version>26.51.0</version>
+        <type>pom</type>
+        <scope>import</scope>
+      </dependency>
+    </dependencies>
+  </dependencyManagement>
   <dependencies>
     <dependency>
       <groupId>org.apache.beam</groupId>
@@ -145,6 +156,10 @@
       <artifactId>junit</artifactId>
       <version>4.13.1</version> <!-- can be omitted when using the BOM -->
       <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>com.google.cloud</groupId>
+      <artifactId>google-cloud-aiplatform</artifactId>
     </dependency>
   </dependencies>
 </project>
