+++ Ch08/src/main/java/org/example/CreateTrainingDataset10.java
@@ -33,7 +33,6 @@ import org.apache.beam.sdk.transforms.windowing.IntervalWindow;
 import org.apache.beam.sdk.transforms.windowing.SlidingWindows;
 import org.apache.beam.sdk.transforms.windowing.Window;
 import org.apache.beam.sdk.values.*;
-import org.example.entities.Flight;
 import org.example.transforms.GroupAndCombine;
 import org.example.transforms.ParDoWithSideInput;
 import org.example.transforms.ParsingIntoObjects;
@@ -46,7 +45,7 @@ import org.slf4j.LoggerFactory;
 import java.util.Arrays;
 import java.util.Map;
 
-import static org.example.entities.Flight.INPUTCOLS.*;
+import static org.example.Flight.INPUTCOLS.*;
 
 public class CreateTrainingDataset10 {
   private static final Logger LOG = LoggerFactory.getLogger(CreateTrainingDataset10.class);
