+++ Ch08/src/main/java/org/example/CreateTrainingDataset4.java
@@ -7,7 +7,6 @@ import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
 import org.apache.beam.sdk.transforms.*;
 import org.apache.beam.sdk.values.PCollectionView;
-import org.example.entities.Flight;
 import org.example.transforms.ParDoWithSideInput;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -16,7 +15,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
-import static org.example.entities.Flight.INPUTCOLS.FL_DATE;
+import static org.example.Flight.INPUTCOLS.FL_DATE;
 
 public class CreateTrainingDataset4 {
     private static final Logger LOG = LoggerFactory.getLogger(CreateTrainingDataset4.class);
