+++ Ch08/src/main/java/org/example/CreateTrainingDataset6.java
@@ -21,7 +21,6 @@ import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.transforms.*;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.PCollectionView;
-import org.example.entities.Flight;
 import org.example.transforms.GroupAndCombine;
 import org.example.transforms.MutatingTheFlightObject;
 import org.example.transforms.ParsingIntoObjects;
