+++ Ch08/src/main/java/org/example/CreateTrainingDataset8.java
@@ -33,7 +33,6 @@ import org.apache.beam.sdk.transforms.windowing.Window;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.PCollectionView;
-import org.example.entities.Flight;
 import org.example.transforms.GroupAndCombine;
 import org.example.transforms.MutatingTheFlightObject;
 import org.example.transforms.ParDoWithSideInput;
@@ -44,7 +43,7 @@ import org.slf4j.LoggerFactory;
 
 import java.util.Map;
 
-import static org.example.entities.Flight.INPUTCOLS.*;
+import static org.example.Flight.INPUTCOLS.*;
 
 public class CreateTrainingDataset8 {
   private static final Logger LOG = LoggerFactory.getLogger(CreateTrainingDataset8.class);
