+++ Ch08/src/test/java/org/example/TestGroupAndCombine.java
@@ -8,7 +8,6 @@ import org.apache.beam.sdk.transforms.ParDo;
 import org.apache.beam.sdk.values.KV;
 import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TypeDescriptors;
-import org.example.entities.Flight;
 import org.example.transforms.GroupAndCombine;
 import org.example.transforms.ParsingIntoObjects;
 import org.junit.Before;
@@ -17,7 +16,7 @@ import org.junit.Test;
 
 import java.util.*;
 
-import static org.example.entities.Flight.INPUTCOLS.NOTIFY_TIME;
+import static org.example.Flight.INPUTCOLS.NOTIFY_TIME;
 
 public class TestGroupAndCombine {
     @Rule
