+++ Ch08/src/test/java/org/example/TestParDoWithSideInput.java
@@ -7,7 +7,6 @@ import org.apache.beam.sdk.transforms.MapElements;
 import org.apache.beam.sdk.transforms.ParDo;
 import org.apache.beam.sdk.transforms.View;
 import org.apache.beam.sdk.values.*;
-import org.example.entities.Flight;
 import org.example.transforms.ParDoWithSideInput;
 import org.junit.Before;
 import org.junit.Rule;
@@ -17,7 +16,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import static org.example.entities.Flight.INPUTCOLS.NOTIFY_TIME;
+import static org.example.Flight.INPUTCOLS.NOTIFY_TIME;
 
 
 public class TestParDoWithSideInput {
